requirements.txt
streamlit
pdfplumber
python-docx
openai